requirements.txt
# ---- Basic ----
numpy>=1.23
Pillow>=9.4
opencv-python>=4.8
tqdm>=4.66
scikit-image>=0.21
scipy>=1.10

# ---- Config / Utils ----
omegaconf>=2.3
einops>=0.6

# ---- Diffusion / HF ----
diffusers>=0.25.0
transformers>=4.37.0
accelerate>=1.0.0
huggingface-hub>=0.25.0
safetensors>=0.4.0
ftfy>=6.1
regex>=2022.10.31
timm>=1.0.0

# ---- Metrics ----
lpips>=0.1.4
pyiqa>=0.1.14
torch-fidelity>=0.3.0

# ---- Data / Log ----
datasets>=3.0.0          
wandb>=0.16.0             
gdown>=4.4.0              

# ---- Model ----
realesrgan>=0.3.0
basicsr>=1.4.2
facexlib>=0.3.0
